chore: rename folder for managed services api
