+++ .golangci.yaml
@@ -13,7 +13,7 @@ run:
   # "/" will be replaced by current OS file path separator to properly work
   # on Windows.
   skip-dirs:
-    - pkg/api/managedservices/client
+    - pkg/api/managedservices
 # output configuration options
 output:
   # colored-line-number|line-number|json|tab|checkstyle|code-climate, default is "colored-line-number"
