+++ cmd/rhoas/kafka/get/get.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	ms "github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/managedservices"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/builders"
 	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/kafka"
 )
 
@@ -36,7 +36,7 @@ func runGet(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	client := ms.BuildClient()
+	client := builders.BuildClient()
 
 	response, status, err := client.DefaultApi.GetKafkaById(context.Background(), id)
