+++ cmd/rhoas/kafka/use/use.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	ms "github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/managedservices"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/builders"
 	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/config"
 )
 
@@ -30,7 +30,7 @@ func runUse(cmd *cobra.Command, args []string) {
 
 	id := args[0]
 
-	client := ms.BuildClient()
+	client := builders.BuildClient()
 
 	res, status, err := client.DefaultApi.GetKafkaById(context.Background(), id)
 	if err != nil {
