+++ pkg/api/managedservices/.openapi-generator-ignore
