+++ pkg/api/managedservices/.openapi-generator/VERSION
