+++ pkg/api/managedservices/.travis.yml
