+++ pkg/api/managedservices/api/openapi.yaml
