+++ pkg/api/managedservices/client.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 import (
 	"bytes"
