+++ pkg/api/managedservices/docs/Error.md
