+++ pkg/api/managedservices/docs/KafkaRequest.md
