+++ pkg/api/managedservices/docs/KafkaRequestAllOf.md
