+++ pkg/api/managedservices/docs/KafkaRequestList.md
