+++ pkg/api/managedservices/docs/ObjectReference.md
