+++ pkg/api/managedservices/docs/TokenResponse.md
