+++ pkg/api/managedservices/docs/TokenResponseAllOf.md
