+++ pkg/api/managedservices/git_push.sh
