+++ pkg/api/managedservices/model_error.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // Error struct for Error
 type Error struct {
