+++ pkg/api/managedservices/model_error_all_of.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // ErrorAllOf struct for ErrorAllOf
 type ErrorAllOf struct {
