+++ pkg/api/managedservices/model_error_list.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // ErrorList struct for ErrorList
 type ErrorList struct {
