+++ pkg/api/managedservices/model_error_list_all_of.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // ErrorListAllOf struct for ErrorListAllOf
 type ErrorListAllOf struct {
