+++ pkg/api/managedservices/model_kafka_request.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 import (
 	"time"
