+++ pkg/api/managedservices/model_kafka_request_all_of.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 import (
 	"time"
