+++ pkg/api/managedservices/model_kafka_request_list.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // KafkaRequestList struct for KafkaRequestList
 type KafkaRequestList struct {
