+++ pkg/api/managedservices/model_list.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // List struct for List
 type List struct {
