+++ pkg/api/managedservices/model_object_reference.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 // ObjectReference struct for ObjectReference
 type ObjectReference struct {
