+++ pkg/api/managedservices/model_token_response.go
@@ -7,10 +7,13 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
-// TokenResponseAllOf struct for TokenResponseAllOf
-type TokenResponseAllOf struct {
+// TokenResponse struct for TokenResponse
+type TokenResponse struct {
+	Id           string `json:"id,omitempty"`
+	Kind         string `json:"kind,omitempty"`
+	Href         string `json:"href,omitempty"`
 	ClientID     string `json:"clientID,omitempty"`
 	ClientSecret string `json:"clientSecret,omitempty"`
 }
