+++ pkg/api/managedservices/model_token_response_all_of.go
@@ -7,13 +7,10 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
-// TokenResponse struct for TokenResponse
-type TokenResponse struct {
-	Id           string `json:"id,omitempty"`
-	Kind         string `json:"kind,omitempty"`
-	Href         string `json:"href,omitempty"`
+// TokenResponseAllOf struct for TokenResponseAllOf
+type TokenResponseAllOf struct {
 	ClientID     string `json:"clientID,omitempty"`
 	ClientSecret string `json:"clientSecret,omitempty"`
 }
