+++ pkg/api/managedservices/response.go
@@ -7,7 +7,7 @@
  * Generated by: OpenAPI Generator (https://openapi-generator.tech)
  */
 
-package msapi
+package managedservices
 
 import (
 	"net/http"
